Eliminated some error messges (#957)
patchuninstall would give error about "zip had nothing to do" if
the patchinstall only added files and directories but did not
modify any existing files.
vnmrpipe would give error is the NMRPpe dynamo application was not
installed.
If a fid is displayed (df) and then makefid makes a new fid, a
subsequent ft would give a message about data file not open.

+++ src/scripts/patchuninstall.sh
@@ -78,12 +78,15 @@ uninstallPatch () {
       source $dir/p_uninstall
       rm p_uninstall
    fi
-   zip -qry patch.zip *
-   mv patch.zip $vnmrsystem
-   cd $vnmrsystem
+   fileList=$(ls)
+   if [[ ! -z $fileList ]]; then
+      zip -qry patch.zip *
+      mv patch.zip $vnmrsystem
+      cd $vnmrsystem
+      unzip -q -o patch.zip
+      rm patch.zip
+   fi
    rm -rf $dir
-   unzip -q -o patch.zip
-   rm patch.zip
    return 0
 }
